moving tfmodels_detectors module up one-level; adding default deployment support
