+++ eta/detectors/__init__.py
@@ -8,5 +8,4 @@ Brian Moore, brian@voxel51.com
 '''
 
 # Import all detectors into the `eta.detectors` namespace
-from .tfmodels.tfmodels_detectors import TFModelsDetector, \
-                                         TFModelsDetectorConfig
+from .tfmodels_detectors import TFModelsDetector, TFModelsDetectorConfig
